rename: ArtifactLoader -> AssemblerInfoProvider. move: Move getContractDataIfExists to AssemblerInfoProvider class.
